Fixes bug where one-digit servo IDs are casted into char
Issue arises when one-digit servo IDs are specified in the YAML file. Since
Protocol::id_t is uint8_t (a.k.a. char), the stream insertion operator
converts the ID into a char. This results in mappings that are not consistent
with the settings in the YAML file. For example,

hardware_corrections:
  3: 3.14

is interpreted as:

hardware_corrections:
  51: 3.14

because the ASCII code for '3' is 51.